pyproject.toml
[tool.poetry]
name = "fastapi-jwt-auth-template"
version = "0.1.0"
description = "FastAPI JWT Authentication Template"
authors = ["Dennis Lee (dennislwy) <wylee2000@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
aiosqlite = "^0.19.0"
pydantic-settings = "^2.1.0"
aiocache = "^0.12.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.25"
python-multipart = "^0.0.7"
email-validator = "^2.1.0.post1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
